refactor: move helpers.ts from github/ to api/ and change imports (#8)
